backend-v2/pyproject.toml
[project]
name = "backend-v2"
version = "0.1.0"
description = "V2 backend of quibble"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-allauth[headless-spec,socialaccount]>=65.11.1",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-ninja>=1.4.3",
    "django-resized>=1.0.3",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.2",
]

[tool.pyright]
reportAny = "none"
reportExplicitAny = "none"
reportUnusedCallResult = "none"
reportUnknownVariableType = "none"
reportUnknownMemberType = "none"
reportUnannotatedClassAttribute = "none"
reportMissingTypeStubs = "none"
